refactor(packaging): move 8 residual flat modules into larch.* packages
Moves ctx, env_file, coder_delta_guards into larch.core; design_diagram_log,
exec_issue_detail, review_phase_detail, run_log_tolerance into larch.report;
and self_review_tally into larch.review. Rewrites all 14 runtime import sites
and all test imports. Deletes design_legacy.py (no runtime importers).

Closes #5655
