+++ python/env-via-config-constant-baseline.json
@@ -1,23 +1,5 @@
 [
   {
-    "file": "design_legacy.py",
-    "qualified_symbol": "_script_root_for",
-    "env_name": "CLAUDE_PLUGIN_ROOT",
-    "constant": "ENV_CLAUDE_PLUGIN_ROOT",
-    "access": "get",
-    "occurrence": 1,
-    "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
-  },
-  {
-    "file": "exec_issue_detail.py",
-    "qualified_symbol": "_plugin_root",
-    "env_name": "CLAUDE_PLUGIN_ROOT",
-    "constant": "ENV_CLAUDE_PLUGIN_ROOT",
-    "access": "get",
-    "occurrence": 1,
-    "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
-  },
-  {
     "file": "larch/agents/agent_voters.py",
     "qualified_symbol": "_append_voter1_failure",
     "env_name": "IMPLEMENT_TMPDIR",
