+++ python/larch/design/plan_quality.py
@@ -26,7 +26,7 @@ from collections.abc import Callable, Iterable
 from larch.agents import agents
 from larch.core import config
 from larch.core import external_defaults
-from ctx import Ctx
+from larch.core.ctx import Ctx
 from larch.design import design_pause
 from larch.issue.issue_wire import emit_untrusted_file_block
 from larch.core.logging_util import diagnostic, emit, emit_kv, quiet_init, reset_quiet_state
