+++ python/larch/report/design_diagram_log.py
