+++ python/larch/report/progress_report.py
@@ -17,7 +17,7 @@ from pathlib import Path
 from typing import cast
 
 from larch.agents import collect_results
-import env_file
+from larch.core import env_file
 from larch.rendering.gantt import GanttRow, format_mss, render_gantt
 from larch import io as larch_io
 from larch.core import config
