+++ python/subprocess-via-runner-baseline.json
@@ -1,19 +1,5 @@
 [
   {
-    "file": "design_legacy.py",
-    "qualified_symbol": "run_script",
-    "callee": "run",
-    "occurrence": 1,
-    "reason": "grandfathered direct subprocess usage pre-G-Py-9 ratchet"
-  },
-  {
-    "file": "exec_issue_detail.py",
-    "qualified_symbol": "assess_issue_details",
-    "callee": "run",
-    "occurrence": 1,
-    "reason": "grandfathered direct subprocess usage pre-G-Py-9 ratchet"
-  },
-  {
     "file": "larch/agents/agent_voters.py",
     "qualified_symbol": "_launch_claude_voter",
     "callee": "Popen",
